Extract segment submission, voting, and UI component code into content/segmentSubmission.ts
Extract 33 functions from content.ts into a new content/segmentSubmission.ts module:
- Segment creation: startOrEndTimingNewSegment, getIncompleteSegment,
  isSegmentCreationInProgress, cancelCreatingSegment
- Segment submission: updateSponsorTimesSubmitting, checkForPreloadedSegment,
  submitSegments, sendSubmitMessage, getSegmentsMessage
- Voting: vote, voteAsync
- UI setup: setupSkipButtonControlBar, setupCategoryPill, setupDescriptionPill,
  updateVisibilityOfPlayerControlsButton, updateSegmentSubmitting
- Info/submission menus: openInfoMenu, closeInfoMenu, openSubmissionMenu,
  closeSubmissionMenu, resetSponsorSubmissionNotice
- Port video: updatePortVideoElements, getPortVideo, submitPortVideo,
  portVideoVote, updateSegments
- Server lookups: sponsorsLookup, lockedCategoriesLookup
- Notices: closeAllSkipNotices, dontShowNoticeAgain
- Misc: getRealCurrentTime, clearSponsorTimes, previewRecentSegment

Uses dependency injection (initSegmentSubmission) for cross-module deps:
skipToTime, startSponsorSchedule, previewTime,
startSkipScheduleCheckingForStartSponsors, updatePreviewBar, selectSegment,
seekFrameByKeyPressListener, playerButton, skipNoticeContentContainer.
